Move RecordRetrievalException to query-core
