+++ java/query/query-core/src/main/java/sleeper/query/core/RecordRetrievalException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.runner.recordretrieval;
+package sleeper.query.core;
 
 public class RecordRetrievalException extends Exception {
