+++ java/query/query-runner/src/main/java/sleeper/query/runner/recordretrieval/LeafPartitionRecordRetriever.java
@@ -18,6 +18,7 @@ package sleeper.query.runner.recordretrieval;
 import sleeper.core.iterator.CloseableIterator;
 import sleeper.core.record.Record;
 import sleeper.core.schema.Schema;
+import sleeper.query.core.RecordRetrievalException;
 import sleeper.query.core.model.LeafPartitionQuery;
 
 @FunctionalInterface
