+++ java/query/query-runner/src/test/java/sleeper/query/runner/recordretrieval/QueryExecutorTest.java
@@ -39,6 +39,7 @@ import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.FileReferenceFactory;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
+import sleeper.query.core.RecordRetrievalException;
 import sleeper.query.core.model.Query;
 import sleeper.query.core.model.QueryException;
 import sleeper.query.core.model.QueryProcessingConfig;
